chore: move the feature gate from factory to scraper
