+++ receiver/apachereceiver/scraper_test.go
@@ -27,6 +27,7 @@ import (
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/config/configtls"
+	"go.opentelemetry.io/collector/featuregate"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/scrapertest"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/scrapertest/golden"
@@ -37,13 +38,16 @@ func TestScraper(t *testing.T) {
 	apacheMock := newMockServer(t)
 	cfg := createDefaultConfig().(*Config)
 	cfg.Endpoint = fmt.Sprintf("%s%s", apacheMock.URL, "/server-status?auto")
-	// Let this test check if it works with the feature enabled and the integration test will test the feature disabled.
-	cfg.emitServerNameAsResourceAttribute = true
 	require.NoError(t, cfg.Validate())
+
+	// Let this test check if it works with the feature enabled and the integration test will test the feature disabled.
+	err := featuregate.GetRegistry().Apply(map[string]bool{EmitServerNameAsResourceAttribute: true})
+
+	require.NoError(t, err)
 
 	scraper := newApacheScraper(componenttest.NewNopReceiverCreateSettings(), cfg)
 
-	err := scraper.start(context.Background(), componenttest.NewNopHost())
+	err = scraper.start(context.Background(), componenttest.NewNopHost())
 	require.NoError(t, err)
 
 	actualMetrics, err := scraper.scrape(context.Background())
